+++ src/org/japo/java/events/MWEM.java
@@ -15,33 +15,27 @@
  */
 package org.japo.java.events;
 
-import java.awt.event.FocusEvent;
-import java.awt.event.FocusListener;
+import java.awt.event.MouseWheelEvent;
+import java.awt.event.MouseWheelListener;
 import org.japo.java.forms.GUI;
 
 /**
  *
  * @author José A. Pacheco Ondoño - joanpaon@gmail.com
  */
-public class FEM implements FocusListener {
+public class MWEM implements MouseWheelListener {
 
     // Referencia al igu
     private final GUI gui;
 
-    public FEM(GUI gui) {
+    public MWEM(GUI gui) {
         this.gui = gui;
     }
 
     @Override
-    public void focusGained(FocusEvent e) {
-        // Evento de Foco - GANADO
-        // gui.procesarFocoGanado(e);
-    }
-
-    @Override
-    public void focusLost(FocusEvent e) {
-        // Evento de Foco - PERDIDO
-        // gui.procesarFocoPerdido(e);
+    public void mouseWheelMoved(MouseWheelEvent e) {
+        // Evento de Raton - Movimiento Rueda
+        // gui.procesarMovimientoRueda();
     }
 
 }